feat: update sample app
